PFS-143 Move CSV header validation to back-end
