+++ finance-admin/internal/server/get_uploads.go
@@ -1,12 +1,12 @@
 package server
 
 import (
-	"github.com/opg-sirius-finance-admin/finance-admin/internal/model"
+	"github.com/opg-sirius-finance-admin/shared"
 	"net/http"
 )
 
 type GetUploadsVars struct {
-	ReportsUploadTypes *[]model.ReportUploadType
+	ReportsUploadTypes *[]shared.ReportUploadType
 	AppVars
 }
 
@@ -15,7 +15,7 @@ type GetUploadsHandler struct {
 }
 
 func (h *GetUploadsHandler) render(v AppVars, w http.ResponseWriter, r *http.Request) error {
-	data := GetUploadsVars{&model.ReportUploadTypes, v}
+	data := GetUploadsVars{&shared.ReportUploadTypes, v}
 	data.selectTab("uploads")
 	return h.execute(w, r, data)
 }
