+++ finance-admin/internal/server/upload.go
@@ -1,17 +1,12 @@
 package server
 
 import (
-	"encoding/csv"
 	"errors"
 	"fmt"
 	"github.com/opg-sirius-finance-admin/finance-admin/internal/api"
 	"github.com/opg-sirius-finance-admin/finance-admin/internal/model"
 	"github.com/opg-sirius-finance-admin/shared"
-	"io"
 	"net/http"
-	"reflect"
-	"strings"
-	"unicode"
 )
 
 type UploadHandler struct {
@@ -32,28 +27,6 @@ func (h *UploadHandler) render(v AppVars, w http.ResponseWriter, r *http.Request
 	}
 	defer file.Close()
 
-	csvReader := csv.NewReader(file)
-	expectedHeaders := reportHeadersByType(reportUploadType)
-
-	readHeaders, err := csvReader.Read()
-	if err != nil {
-		return h.handleError(w, r, "Failed to read CSV headers", http.StatusBadRequest)
-	}
-
-	for i, header := range readHeaders {
-		readHeaders[i] = cleanString(header)
-	}
-
-	// Compare the extracted headers with the expected headers
-	if !reflect.DeepEqual(readHeaders, expectedHeaders) {
-		return h.handleError(w, r, "CSV headers do not match for the file trying to be uploaded", http.StatusBadRequest)
-	}
-
-	_, err = file.Seek(0, io.SeekStart)
-	if err != nil {
-		return err
-	}
-
 	data, err := shared.NewUpload(reportUploadType, uploadDate, email, file, handler.Filename)
 	if err != nil {
 		return h.handleError(w, r, "Failed to read file", http.StatusBadRequest)
@@ -67,19 +40,6 @@ func (h *UploadHandler) render(v AppVars, w http.ResponseWriter, r *http.Request
 	w.Header().Add("HX-Redirect", fmt.Sprintf("%s/uploads?success=upload", v.EnvironmentVars.Prefix))
 
 	return nil
-}
-
-func reportHeadersByType(reportType string) []string {
-	switch reportType {
-	case model.ReportTypeUploadDeputySchedule.Key():
-		return []string{"Deputy number", "Deputy name", "Case number", "Client forename", "Client surname", "Do not invoice", "Total outstanding"}
-	case model.ReportTypeUploadDebtChase.Key():
-		return []string{"Client_no", "Deputy_name", "Total_debt"}
-	case model.ReportTypeUploadPaymentsOPGBACS.Key():
-		return []string{"Line", "Type", "Code", "Number", "Transaction", "Value Date", "Amount", "Amount Reconciled", "Charges", "Status", "Desc Flex", "Consolidated line"}
-	default:
-		return []string{"Unknown report type"}
-	}
 }
 
 // handleError simplifies repetitive error handling in the render method.
@@ -109,16 +69,4 @@ func (h *UploadHandler) handleUploadError(w http.ResponseWriter, r *http.Request
 	}
 
 	return err
-}
-
-func cleanString(s string) string {
-	// Trim leading and trailing spaces
-	s = strings.TrimSpace(s)
-	// Remove non-printable characters
-	return strings.Map(func(r rune) rune {
-		if unicode.IsPrint(r) {
-			return r
-		}
-		return -1
-	}, s)
 }
