+++ shared/upload_type.go
@@ -1,4 +1,4 @@
-package model
+package shared
 
 import "encoding/json"
